Cleanup and split some js scripts
