+++ oioioi/base/static/js/bootstrap-tooltip-setup.js
@@ -1,0 +1,3 @@
+$(function () {
+    $("[data-toggle='tooltip']").tooltip();
+});
