+++ oioioi/base/static/js/highlight-setup.js
@@ -1,0 +1,20 @@
+import hljs from 'highlight.js/lib/core';
+
+// Selection of common languages to limit bundle size
+import javascript from 'highlight.js/lib/languages/javascript';
+import python from 'highlight.js/lib/languages/python';
+import java from 'highlight.js/lib/languages/java';
+import bash from 'highlight.js/lib/languages/bash';
+import xml from 'highlight.js/lib/languages/xml';
+import cpp from 'highlight.js/lib/languages/cpp';
+import c from 'highlight.js/lib/languages/c';
+
+hljs.registerLanguage('javascript', javascript);
+hljs.registerLanguage('python', python);
+hljs.registerLanguage('java', java);
+hljs.registerLanguage('bash', bash);
+hljs.registerLanguage('xml', xml);
+hljs.registerLanguage('cpp', cpp);
+hljs.registerLanguage('c', c);
+
+hljs.highlightAll()
