--- oioioi/base/static/js/highlight.js
@@ -1,11 +1,0 @@
-import hljs from 'highlight.js/lib/core';
-
-import c from 'highlight.js/lib/languages/c';
-import cpp from 'highlight.js/lib/languages/cpp';
-import python from 'highlight.js/lib/languages/python';
-
-hljs.registerLanguage('c', c);
-hljs.registerLanguage('cpp', cpp);
-hljs.registerLanguage('python', python);
-
-hljs.highlightAll()
