+++ oioioi/base/static/js/menu-setup.js
